You can now change text on the fly! (also organized stuff)
